refactorizacion

+++ app/Http/Controllers/AlumnoCursoController.php
@@ -10,13 +10,13 @@ use Intranet\Entities\Alumno_grupo;
 use Intranet\Entities\Curso;
 use Intranet\Botones\BotonImg;
 use Illuminate\Support\Facades\Session;
+use Intranet\Entities\Alumno;
 
 class AlumnoCursoController extends IntranetController
 {
 
     use traitImprimir;
 
-    
     protected $model = 'AlumnoCurso';
     protected $gridFields = ['idAlumno', 'nombre', 'finalizado', 'registrado'];
 
@@ -24,14 +24,11 @@ class AlumnoCursoController extends IntranetController
     {
         return AlumnoCurso::where('idCurso', '=', $this->search)->get();
     }
-    
+
     public function active($id)
     {
         $actual = AlumnoCurso::find($id);
-        if ($actual->finalizado)
-            $actual->finalizado = 0;
-        else
-            $actual->finalizado = 1;
+        $actual->finalizado = $actual->finalizado ? 0 : 1;
         $actual->save();
         return back();
     }
@@ -50,49 +47,15 @@ class AlumnoCursoController extends IntranetController
         $this->panel->setBoton('grid', new BotonImg('alumnocurso.delete', ['where' => ['finalizado', '==', 0]]));
         $this->panel->setBoton('grid', new BotonImg('alumnocurso.pdf', ['where' => ['finalizado', '==', 1]]));
     }
+
     public function pdf($id)
     {
-        $actual = AlumnoCurso::where('id',$id)->get();
+        $actual = AlumnoCurso::where('id', $id)->get();
         $curso = Curso::find($actual->first()->idCurso);
-        if (haVencido($curso->fecha_fin)){
-            return self::hazPdf('pdf.alumnos.manipulador',$actual,$curso)->stream();
-        }
-        else return self::imprime($id);
-    }
-
-    public function register($id, $alumno = null, $redirect = 1)
-    {
-        $curso = Curso::find($id);
-        $alumno = $alumno ? $alumno : AuthUser()->nia;
-        if ($curso) {
-            $existe = AlumnoCurso::where('idCurso', $id)
-                    ->where('idAlumno', $alumno)
-                    ->count();
-            if (!$existe) {
-                $new = new AlumnoCurso();
-                $new->idAlumno = $alumno;
-                $new->idCurso = $id;
-                $new->finalizado = 0;
-                if ($curso->aforo == 0)
-                    $new->registrado = 'S';
-                else {
-                    if ($curso->NAlumnos < $curso->aforo)
-                        $new->registrado = 'S';
-                    else {
-                        if ($curso->NAlumnos < $curso->aforo * config('variables.reservaAforo'))
-                            $new->registrado = 'R';
-                        else
-                            $new->registrado = 'N';
-                    }
-                }
-                if ($new->registrado != 'N')
-                    $new->save();
-                avisa($alumno, $this->mensaje($new->registrado));
-            }
-        }
-
-        if ($redirect)
-            return back();
+        if (haVencido($curso->fecha_fin)) {
+            return self::hazPdf('pdf.alumnos.manipulador', $actual, $curso)->stream();
+        } else
+            return self::imprime($id);
     }
 
     public function registerG($grupo, $id)
@@ -109,51 +72,56 @@ class AlumnoCursoController extends IntranetController
         return $this->register($id, $alumno);
     }
 
-    public function unregister($id, $alumno = null, $redirect = 1)
+    public function register($id, $alumno = null, $redirect = 1)
     {
         $curso = Curso::find($id);
-        $alumno = $alumno ? $alumno : AuthUser()->nia;
-
-        if ($curso) {
-            $existe = AlumnoCurso::where('idCurso', $id)
-                    ->where('idAlumno', $alumno)
-                    ->count();
-            if ($existe) {
-                $registro = AlumnoCurso::where('idCurso', $id)
-                        ->where('idAlumno', $alumno)
-                        ->first();
-                $tipoRegistro = $registro->registrado;
-                $registro->delete();
-                if ($curso->aforo) {
-                    $registro = AlumnoCurso::where('idCurso', $id)
-                            ->where('registrado', 'R')
-                            ->orderBy('id')
-                            ->first();
-                    if ($registro) {
-                        $registro->registrado = 'S';
-                        $registro->save();
-                        avisa($registro->idAlumno, $this->mensaje('S'));
-                    }
-                }
-                avisa($alumno, $this->mensaje('E'));
-            }
+        $alumno = $alumno ? Alumno::find($alumno) : AuthUser();
+        $existe = AlumnoCurso::where('idCurso', $id)
+                ->where('idAlumno', $alumno->nia)
+                ->count();
+        if ($curso && !$existe) {
+            if ($curso->aforo == 0 || $curso->NAlumnos < $curso->aforo) {
+                $alumno->Curso()->attach($id, ['registrado' => 'S', 'finalizado' => 0]);
+                $mensaje = 'Has sigut registrat/ada al curs ';
+            } else
+            if ($curso->NAlumnos < $curso->aforo * config('variables.reservaAforo')) {
+                $alumno->Curso()->attach($id, ['registrado' => 'R', 'finalizado' => 0]);
+                $mensaje = "Estas en llista d'espera al curs ";
+            } else
+                $mensaje = 'No queden places al curs ';
+            avisa($alumno->nia, $mensaje);
         }
         if ($redirect)
             return back();
     }
 
-    private function mensaje($tipoMensaje)
+    public function unregister($id, $alumno = null, $redirect = 1)
     {
-        switch ($tipoMensaje) {
-            case 'S' : $mensaje = 'Has sigut registrat/ada al curs ';
-                break;
-            case 'R' : $mensaje = "Estas en llista d'espera al curs ";
-                break;
-            case 'N' : $mensaje = 'No queden places al curs ';
-                break;
-            case 'E' : $mensaje = 'Esborrat/ada del curs ';
-        }
-        return $mensaje;
-    }
+        $curso = Curso::find($id);
+        $alumno = $alumno ? $alumno : AuthUser()->nia;
+        $existe = AlumnoCurso::where('idCurso', $id)
+                ->where('idAlumno', $alumno)
+                ->count();
 
+        if ($curso && $existe) {
+            $registro = AlumnoCurso::where('idCurso', $id)
+                    ->where('idAlumno', $alumno)
+                    ->first();
+            $tipoRegistro = $registro->registrado;
+            $registro->delete();
+            if ($curso->aforo && $tipoRegistro == 'S') {
+                $registro = AlumnoCurso::where('idCurso', $id)
+                        ->where('registrado', 'R')
+                        ->orderBy('id')
+                        ->first();
+                if ($registro) {
+                    $registro->registrado = 'S';
+                    $registro->save();
+                    avisa($registro->idAlumno, 'Has sigut registrat/ada al curs ');
+                }
+            }
+            avisa($alumno, 'Esborrat/ada del curs ');
+        }
+        if ($redirect) return back();
+    }
 }
